Move object store credential parsing into object store module
